refactor: 프로젝션용 DTO 패키지 위치 변경
