+++ backend/bottari/src/main/java/com/bottari/teambottari/service/TeamBottariService.java
@@ -12,7 +12,7 @@ import com.bottari.teambottari.domain.TeamPersonalItem;
 import com.bottari.teambottari.domain.TeamSharedItem;
 import com.bottari.teambottari.dto.CreateTeamBottariRequest;
 import com.bottari.teambottari.dto.ReadTeamBottariPreviewResponse;
-import com.bottari.teambottari.dto.TeamBottariMemberCountProjection;
+import com.bottari.teambottari.repository.dto.TeamBottariMemberCountProjection;
 import com.bottari.teambottari.repository.TeamAssignedItemRepository;
 import com.bottari.teambottari.repository.TeamBottariRepository;
 import com.bottari.teambottari.repository.TeamMemberRepository;
